Move matrix_transpose, matrix_product to matrix_utilities.
