+++ astropy/coordinates/builtin_frames/galactic_transforms.py
@@ -3,14 +3,14 @@
 from __future__ import (absolute_import, unicode_literals, division,
                         print_function)
 
-from ..matrix_utilities import rotation_matrix
+from ..matrix_utilities import (rotation_matrix,
+                                matrix_product, matrix_transpose)
 from ..baseframe import frame_transform_graph
 from ..transformations import DynamicMatrixTransform
 
 from .fk5 import FK5
 from .fk4 import FK4NoETerms
-from .utils import (EQUINOX_B1950, EQUINOX_J2000,
-                    matrix_product, matrix_transpose)
+from .utils import EQUINOX_B1950, EQUINOX_J2000
 from .galactic import Galactic
