+++ astropy/coordinates/earth_orientation.py
@@ -15,8 +15,7 @@ import numpy as np
 
 from ..time import Time
 from .. import units as u
-from .matrix_utilities import rotation_matrix
-from .builtin_frames.utils import matrix_product, matrix_transpose
+from .matrix_utilities import rotation_matrix, matrix_product, matrix_transpose
 
 
 jd1950 = Time('B1950', scale='tai').jd
